use correct types

+++ firedrake/evaluate.h
@@ -2,6 +2,7 @@
 #define _EVALUATE_H
 
 #include <petsc.h>
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -54,7 +55,7 @@ extern int locate_cell_from_candidates(struct Function *f,
 		       void *found_ref_coords,
 		       double *found_ref_cell_dist_l1,
 			   size_t nids,
-			   size_t *ids,
+			   int64_t *ids,
 			   size_t ncells_ignore,
 			   int* cells_ignore);
